Get rid of pepesort and maybesort
Pepesort:

pepesort was an in-place mergesort that i had developed myself.
Originally, it ran in a time complexity greater than O(NlogN), but less
than O(N^2). I do not remember what it exactly was.

Over time, I did some small and some big changes to it, and it ended up
converging onto a traditional in place mergesort. For this reason, I
consider it is not interesting to have it here anymore.

Maybesort:

Maybesort was meant as a joke. It is an intentionally broken version of
quicksort designed to introduce randomized algorithms.